// frontend/generated/models/AuthResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AccountDto } from './AccountDto';
import {
    AccountDtoFromJSON,
    AccountDtoFromJSONTyped,
    AccountDtoToJSON,
    AccountDtoToJSONTyped,
} from './AccountDto';

/**
 * 
 * @export
 * @interface AuthResponseDto
 */
export interface AuthResponseDto {
    /**
     * 
     * @type {string}
     * @memberof AuthResponseDto
     */
    token?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthResponseDto
     */
    refreshToken?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthResponseDto
     */
    email?: string;
    /**
     * 
     * @type {AccountDto}
     * @memberof AuthResponseDto
     */
    accountResponseDTO?: AccountDto;
}

/**
 * Check if a given object implements the AuthResponseDto interface.
 */
export function instanceOfAuthResponseDto(value: object): value is AuthResponseDto {
    return true;
}

export function AuthResponseDtoFromJSON(json: any): AuthResponseDto {
    return AuthResponseDtoFromJSONTyped(json, false);
}

export function AuthResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'token': json['token'] == null ? undefined : json['token'],
        'refreshToken': json['refreshToken'] == null ? undefined : json['refreshToken'],
        'email': json['email'] == null ? undefined : json['email'],
        'accountResponseDTO': json['accountResponseDTO'] == null ? undefined : AccountDtoFromJSON(json['accountResponseDTO']),
    };
}

export function AuthResponseDtoToJSON(json: any): AuthResponseDto {
    return AuthResponseDtoToJSONTyped(json, false);
}

export function AuthResponseDtoToJSONTyped(value?: AuthResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'token': value['token'],
        'refreshToken': value['refreshToken'],
        'email': value['email'],
        'accountResponseDTO': AccountDtoToJSON(value['accountResponseDTO']),
    };
}