// frontend/generated/models/ResetPasswordRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ResetPasswordRequestDto
 */
export interface ResetPasswordRequestDto {
    /**
     * 
     * @type {string}
     * @memberof ResetPasswordRequestDto
     */
    token?: string;
    /**
     * 
     * @type {string}
     * @memberof ResetPasswordRequestDto
     */
    newPassword?: string;
    /**
     * 
     * @type {string}
     * @memberof ResetPasswordRequestDto
     */
    confirmPassword?: string;
}

/**
 * Check if a given object implements the ResetPasswordRequestDto interface.
 */
export function instanceOfResetPasswordRequestDto(value: object): value is ResetPasswordRequestDto {
    return true;
}

export function ResetPasswordRequestDtoFromJSON(json: any): ResetPasswordRequestDto {
    return ResetPasswordRequestDtoFromJSONTyped(json, false);
}

export function ResetPasswordRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResetPasswordRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'token': json['token'] == null ? undefined : json['token'],
        'newPassword': json['newPassword'] == null ? undefined : json['newPassword'],
        'confirmPassword': json['confirmPassword'] == null ? undefined : json['confirmPassword'],
    };
}

export function ResetPasswordRequestDtoToJSON(json: any): ResetPasswordRequestDto {
    return ResetPasswordRequestDtoToJSONTyped(json, false);
}

export function ResetPasswordRequestDtoToJSONTyped(value?: ResetPasswordRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'token': value['token'],
        'newPassword': value['newPassword'],
        'confirmPassword': value['confirmPassword'],
    };
}